roll load-virtual up into build-ideal-tree
